Adds Vampirism. Ported full circle back to here from Hardlight.
