+++ Content.Shared/Body/Components/StomachComponent.cs
@@ -58,6 +58,7 @@ namespace Content.Shared.Body.Components
         ///     A whitelist for what special-digestible-required foods this stomach is capable of eating.
         /// </summary>
         [DataField]
+        [Access(Other = AccessPermissions.ReadWriteExecute)]
         public EntityWhitelist? SpecialDigestible = null;
 
         /// <summary>
