+++ Content.Shared/Vampiric/BloodSuckDoAfterEvent.cs
@@ -1,0 +1,10 @@
+using Robust.Shared.Serialization;
+using Content.Shared.DoAfter;
+
+namespace Content.Shared.Vampiric
+{
+
+    public sealed partial class BloodSuckDoAfterEvent : SimpleDoAfterEvent
+    {
+    }
+}
