+++ Resources/Locale/en-US/_Floof/traits/physical_traits.ftl
@@ -74,3 +74,8 @@ trait-heatintolerance-desc =
 trait-heatintoleranceadvanced-name = Heat Intolerance - Advanced
 trait-heatintoleranceadvanced-desc = 
     You simply can't handle the heat. 
+
+trait-name-Vampirism = Vampirism
+trait-description-Vampirism =
+    Your body has evolved to be able to suck blood from beings that contain it and metabolize it into useful compounds.
+    You cannot eat normal food, but drinking blood satiates your hunger and thirst, and also improves your health.
