+++ Resources/Prototypes/_Floof/Traits/physical.yml
@@ -517,3 +517,23 @@
   effects:
   - !type:ModifyTemperatureToleranceEffect
     heatToleranceModifier: -10 # 10 worse (less tolerant) than base (which should be 325).
+
+- type: trait
+  id: Vampirism
+  name: trait-name-Vampirism
+  description: trait-description-Vampirism
+  category: Skills
+  cost: 4 # Whilst they can still eat everything this is literally only a positive trait.
+  speciesBlacklist:
+    - IPC
+    - Synth
+  components:
+    - type: Vampirism
+      succDelay: 3
+      specialDigestible: # Vampires cannot eat food chat is that real
+        tags:
+        - IceCream
+        - Pill
+        - Crayon
+        - Paper
+      isSpecialDigestibleExclusive: true
